fix: use local configuration
When updating PHP, I discovered that the `php.ini` dropped into the CLI
gets rewritten when PHP updates occur. While `apt-get` provides the
ability to review changes, there's no easy way to _merge_ the changes.

As such, this patch makes a change to the `config` command. Instead of
editing the CLI `php.ini` directly, it instead adds a configuration file
to the `/etc/php/{VERSION}/mods-available/` directory, and runs
`phpenmod` to activate it in the CLI SAPI immediately. Any time `config`
is used, it will open this file in the editor, providing a way to ensure
any configuration changes are propagated even after updates.