feat: adds version:enable command
Creates ~/.local/var/lib/alternatives/php if it does not exist, using a
default template. (If the file does not exist, the user is prompted to
create it, aborting if they do not agree.)

From there, it will append the binary name and version to the file.

If the version is already found, it will indicate this to the user.